Popup Menu: Move icons to the left and submenus to the right

+++ extensions/content/portal/classes/commands/PortalCopy.class.php
@@ -26,41 +26,43 @@ class PortalCopy extends \AbstractCommand implements \IAjaxCommand, \IFrameComma
           $portalCopy = $portalObject->copy();
 
           //remove broken messages
-          foreach ($portalCopy->get_inventory() as $columnObject) {
-              foreach ($columnObject->get_inventory() as $portletObject) {
-                  if ($portletObject->get_attribute("bid:portlet")==="msg" && !($portletObject instanceof \steam_link)) {
+          if(!($portalCopy instanceof \steam_link)){
+            foreach ($portalCopy->get_inventory() as $columnObject) {
+                foreach ($columnObject->get_inventory() as $portletObject) {
+                    if ($portletObject->get_attribute("bid:portlet")==="msg" && !($portletObject instanceof \steam_link)) {
 
-                      //get ids in attrbute
-                      $oldIds = $portletObject->get_attribute("bid:portlet:content");
-                      $newIds = array();
+                        //get ids in attrbute
+                        $oldIds = $portletObject->get_attribute("bid:portlet:content");
+                        $newIds = array();
 
-                      //delete wrong references messages
-                      foreach ($portletObject->get_inventory() as $oldMessageObject) {
-                          $oldMessageObject->delete();
-                      }
+                        //delete wrong references messages
+                        foreach ($portletObject->get_inventory() as $oldMessageObject) {
+                            $oldMessageObject->delete();
+                        }
 
-                      foreach ($oldIds as $messageId) {
-                          //copy to here
-                          //make new id list
-                          $msgObject = \steam_factory::get_object( $GLOBALS["STEAM"]->get_id(), $messageId );
-                          $msgCopy = $msgObject->copy();
-                          $msgCopy->move($portletObject);
-                          $newIds[]=$msgCopy->get_id();
-                          //handle included pics
-                          $pictrueId = $msgObject->get_attribute("bid:portlet:msg:picture_id");
-                          if ($pictrueId!="") {
-                              $pictureObject = \steam_factory::get_object( $GLOBALS["STEAM"]->get_id(), $pictrueId );
-                              $pictuteCopy = $pictureObject->copy();
-                              $pictuteCopy->move($portletObject); //test
-                              $msgCopy->set_attribute("bid:portlet:msg:picture_id",$pictuteCopy->get_id());
-                              $msgCopy->move($portletObject);
-                          }
+                        foreach ($oldIds as $messageId) {
+                            //copy to here
+                            //make new id list
+                            $msgObject = \steam_factory::get_object( $GLOBALS["STEAM"]->get_id(), $messageId );
+                            $msgCopy = $msgObject->copy();
+                            $msgCopy->move($portletObject);
+                            $newIds[]=$msgCopy->get_id();
+                            //handle included pics
+                            $pictrueId = $msgObject->get_attribute("bid:portlet:msg:picture_id");
+                            if ($pictrueId!="") {
+                                $pictureObject = \steam_factory::get_object( $GLOBALS["STEAM"]->get_id(), $pictrueId );
+                                $pictuteCopy = $pictureObject->copy();
+                                $pictuteCopy->move($portletObject); //test
+                                $msgCopy->set_attribute("bid:portlet:msg:picture_id",$pictuteCopy->get_id());
+                                $msgCopy->move($portletObject);
+                            }
 
-                      }
-                      //save in attrubute
-                      $portletObject->set_attribute("bid:portlet:content",$newIds);
-                  }
-              }
+                        }
+                        //save in attrubute
+                        $portletObject->set_attribute("bid:portlet:content",$newIds);
+                    }
+                }
+            }
           }
 
           $portalCopy->move($this->user);
